+++ CHANGELOG.md
@@ -2,13 +2,15 @@
 
 ## Digital Token 0.6.0
 
-This is the changelog for Digital Token version 0.6.0 released on July 7th, 2023.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/digital_token/tags)
+This is the changelog for Digital Token version 0.6.0 released on July 8th, 2023.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/digital_token/tags)
 
 ### Bug Fixes
 
 * Remove checking Phoenix and Ecto json_library configurations.  Ecto no longer configures its `:json_library` in `config.exs` and checking for Phoenix configuration can cause config failures when [building in Docker](https://github.com/elixir-cldr/cldr/issues/208).
 
 * Sort the short names for a token by their length so that the standard short name is the shortest one.
+
+* Fix code that made an implicit assumption about map order.
 
 ### Enhancements